+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/config/BaseWebAppConfiguration.groovy
@@ -2,7 +2,7 @@ package com.ofg.infrastructure.config
 
 import com.ofg.infrastructure.discovery.ServiceDiscoveryConfiguration
 import com.ofg.infrastructure.healthcheck.HealthCheckConfiguration
-import com.ofg.infrastructure.metrics.registry.MetricsConfiguration
+import com.ofg.infrastructure.metrics.config.MetricsConfiguration
 import com.ofg.infrastructure.web.config.WebInfrastructureConfiguration
 import groovy.transform.CompileStatic
 import org.springframework.context.annotation.Configuration