+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/correlationid/CorrelationIdFilter.groovy
@@ -1,5 +1,6 @@
-package com.ofg.infrastructure.web.filter.correlationid
+package com.ofg.infrastructure.web.correlationid
 
+import com.ofg.infrastructure.correlationid.CorrelationIdHolder
 import groovy.transform.TypeChecked
 import groovy.util.logging.Slf4j
 import org.slf4j.MDC
@@ -10,16 +11,16 @@ import javax.servlet.ServletException
 import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
-import static com.ofg.infrastructure.web.filter.correlationid.CorrelationIdHolder.CORRELATION_ID_HEADER
+import static com.ofg.infrastructure.correlationid.CorrelationIdHolder.CORRELATION_ID_HEADER
 import static org.springframework.util.StringUtils.hasText
 
 /**
- * Filter that takes the value of the {@link CorrelationIdHolder#CORRELATION_ID_HEADER} header 
- * from either request or response and sets it in the {@link CorrelationIdHolder}. It also provides
+ * Filter that takes the value of the {@link com.ofg.infrastructure.correlationid.CorrelationIdHolder#CORRELATION_ID_HEADER} header
+ * from either request or response and sets it in the {@link com.ofg.infrastructure.correlationid.CorrelationIdHolder}. It also provides
  * that value in {@link org.apache.log4j.MDC} logging related class so that logger prints the value of
  * correlation id at each log.
  * 
- * @see CorrelationIdHolder
+ * @see com.ofg.infrastructure.correlationid.CorrelationIdHolder
  * @see org.apache.log4j.MDC
  */
 @TypeChecked