+++ micro-infra-spring-base/src/main/groovy/com/ofg/infrastructure/web/logging/RequestLoggingConfiguration.groovy
@@ -1,4 +1,4 @@
-package com.ofg.infrastructure.web.filter.logging
+package com.ofg.infrastructure.web.logging
 
 import groovy.transform.TypeChecked
 import org.springframework.beans.factory.annotation.Value
@@ -8,14 +8,14 @@ import org.springframework.context.annotation.Configuration
 import javax.servlet.Filter
 /**
  * Configuration that registers a bean that will automatically if DEBUG level of logging is set on
- * {@link com.ofg.infrastructure.web.filter.logging.RequestBodyLoggingContextFilter} 
+ * {@link com.ofg.infrastructure.web.logging.RequestBodyLoggingContextFilter}
  * print request body in logs - you can limit its length by setting a property
  * 
- * @see com.ofg.infrastructure.web.filter.logging.RequestBodyLoggingContextFilter
+ * @see com.ofg.infrastructure.web.logging.RequestBodyLoggingContextFilter
  */
 @TypeChecked
 @Configuration
-class RequestFilterConfiguration {
+class RequestLoggingConfiguration {
 
     @Bean
     Filter requestBodyLoggingContextFilter(@Value('${request.payload.logging.maxlength:2000}') int maxPayloadLength) {