+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/base/ConfigurationWithoutServiceDiscovery.groovy
@@ -1,6 +1,6 @@
 package com.ofg.infrastructure.base
 
-import com.ofg.infrastructure.web.filter.FilterConfiguration
+import com.ofg.infrastructure.web.correlationid.CorrelationIdConfiguration
 import com.ofg.infrastructure.web.view.ViewConfiguration
 import groovy.transform.TypeChecked
 import org.springframework.context.annotation.Configuration
@@ -8,7 +8,6 @@ import org.springframework.context.annotation.Import
 
 @TypeChecked
 @Configuration
-@Import([FilterConfiguration,
-        ViewConfiguration])
+@Import([CorrelationIdConfiguration, ViewConfiguration])
 class ConfigurationWithoutServiceDiscovery {
 }