+++ micro-infra-spring-base/src/test/groovy/com/ofg/infrastructure/base/MvcIntegrationSpec.groovy
@@ -1,5 +1,5 @@
 package com.ofg.infrastructure.base
-import com.ofg.infrastructure.web.filter.correlationid.CorrelationIdFilter
+import com.ofg.infrastructure.web.correlationid.CorrelationIdFilter
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.context.web.WebAppConfiguration
 import org.springframework.test.web.servlet.MockMvc